requirements.txt
opencv-python
numpy
paddle
tqdm
setuptools
torchvision
natsort